# api/openapi.yaml
openapi: 3.1.0
info:
  title: CiftPay API
  version: 1.0.0
  description: |
    CiftPay turns M-Pesa payments into KRA eTIMS invoices.

    Conventions:
    - All money values are integer **cents** (`amount_cents`). KES 2,400.00 = 240000.
    - All timestamps are RFC 3339 UTC.
    - Authenticated endpoints use the `ciftpay_session` HttpOnly cookie plus `X-CSRF-Token` on non-GET.
    - The active organisation is the caller's default membership or the `X-Org-Id` header.
    - Errors use the envelope in `#/components/schemas/Error`.
    - Personal data (MSISDN, KRA PIN) is returned masked unless stated otherwise.
  contact:
    name: CiftPay Engineering
servers:
  - url: http://localhost:8080/v1
    description: Local development
  - url: https://api.ciftpay.co.ke/v1
    description: Production
...
security:
  - sessionCookie: []
paths:
...
  /auth/otp/request:
    post:
      tags: [auth]
      operationId: requestOtp
      security: []
      summary: Send a one-time code to a phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [msisdn]
              properties:
                msisdn:
                  $ref: "#/components/schemas/Msisdn"
                locale:
                  $ref: "#/components/schemas/Locale"
      responses:
        "202":
          description: Code sent (or silently not sent for unknown numbers in prod)
          content:
            application/json:
              schema:
                type: object
                required: [expires_in_seconds]
                properties:
                  expires_in_seconds:
                    type: integer
                    example: 300
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /auth/otp/verify:
    post:
      tags: [auth]
      operationId: verifyOtp
      security: []
      summary: Exchange a one-time code for a session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [msisdn, code]
              properties:
                msisdn:
                  $ref: "#/components/schemas/Msisdn"
                code:
                  type: string
                  pattern: "^[0-9]{6}$"
      responses:
        "200":
          description: Session created. Sets `ciftpay_session` cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
  /csrf:
    get:
      tags: [auth]
      operationId: getCsrfToken
      summary: Retrieve the active session CSRF token
      responses:
        "200":
          description: Session CSRF token
          content:
            application/json:
              schema:
                type: object
                required: [csrf_token]
                properties:
                  csrf_token:
                    type: string
        "401":
          $ref: "#/components/responses/Error"
...
  /orgs:
    get:
      tags: [orgs]
      operationId: listOrgs
      summary: Organisations the caller is a member of
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgMembership"
    post:
      tags: [orgs]
      operationId: createOrg
      summary: Create an organisation (business) with its KRA PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgCreate"
      description: |
        The KRA PIN must match `^[AP]\d{9}[A-Z]$` and is then looked up with the fiscal
        provider (iTax PIN checker). A PIN KRA does not know is rejected with `422 pin_unknown`;
        if the checker is unavailable the org is created with `kra_pin_verified_at: null`.
      responses:
        "201":
          description: Created; caller becomes `owner`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        "409":
          description: A business with this KRA PIN already exists (`conflict`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/Error"
        default:
          $ref: "#/components/responses/Error"
...
  /org/members/{userId}:
    delete:
      tags: [org]
      operationId: removeOrgMember
      summary: Remove a member from the organization
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /org/invites:
    get:
      tags: [org]
      operationId: listOrgInvites
      summary: List pending invites for the organization
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgInvite"
    post:
      tags: [org]
      operationId: createOrgInvite
      summary: Invite an accountant or team member via phone or email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgInviteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgInvite"
  /org/invites/{id}:
    delete:
      tags: [org]
      operationId: revokeOrgInvite
      summary: Revoke a pending invite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
  /accountant/invites/{id}/accept:
    post:
      tags: [accountant]
      operationId: acceptAccountantInvite
      summary: Accept a pending client invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /accountant/invites/{id}/reject:
    post:
      tags: [accountant]
      operationId: rejectAccountantInvite
      summary: Reject a pending client invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ciftpay_session
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code, snake_case
              examples: [validation_failed, not_found, shortcode_claimed, rate_limited, unauthenticated, forbidden, conflict, internal]
            message:
              type: string
              description: Human-readable, safe to show to users
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
...
    Msisdn:
      type: string
      description: Kenyan mobile number in E.164 without plus, e.g. 254712345678
      pattern: "^254[17][0-9]{8}$"
      example: "254712345678"
    MaskedMsisdn:
      type: string
      description: First 4 and last 3 digits, e.g. 2547•••••345
      example: "2547•••••345"
    KraPin:
      type: string
      pattern: "^[AP][0-9]{9}[A-Z]$"
      example: "A123456789B"
...
    MaskedKraPin:
      type: string
      example: "A•••••••••B"
    Locale:
      type: string
      enum: [en, sw]
      default: en
    Role:
      type: string
      enum: [owner, staff, accountant, admin]
...
    Session:
      type: object
      required: [user_id, csrf_token, expires_at, orgs]
      properties:
        user_id:
          type: string
          format: uuid
        csrf_token:
          type: string
        expires_at:
          type: string
          format: date-time
        msisdn_masked:
          $ref: "#/components/schemas/MaskedMsisdn"
        orgs:
          type: array
          items:
            $ref: "#/components/schemas/OrgMembership"
    OrgMembership:
      type: object
      required: [org_id, name, role, is_default]
      properties:
        org_id:
          type: string
          format: uuid
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        is_default:
          type: boolean
    OrgCreate:
      type: object
      required: [name, kra_pin]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        kra_pin:
          $ref: "#/components/schemas/KraPin"
        vat_registered:
          type: boolean
          default: false
        locale:
          $ref: "#/components/schemas/Locale"
    Org:
      type: object
      required: [id, name, kra_pin_masked, vat_registered, locale, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        kra_pin:
          $ref: "#/components/schemas/KraPin"
        kra_pin_masked:
          $ref: "#/components/schemas/MaskedKraPin"
        kra_pin_verified_at:
          type: [string, "null"]
          format: date-time
        vat_registered:
          type: boolean
        locale:
          $ref: "#/components/schemas/Locale"
        fiscal_adapter:
          type: string
          enum: [mock, vendor, oscu]
        created_at:
          type: string
          format: date-time
...
    OrgInvite:
      type: object
      required: [id, org_id, role, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    OrgInviteInput:
      type: object
      required: []
      properties:
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [accountant, staff, admin]
          default: accountant
...